split clipping page and select clip page
