+++ packages/app/app/studio/[organization]/clips/[stageId]/Timeline/PlayHead.tsx
