+++ packages/app/app/studio/[organization]/clips/[stageId]/Timeline/TrimmControls.tsx
