+++ packages/app/app/studio/[organization]/clips/[stageId]/sidebar/clips/Clip.tsx
