+++ packages/app/app/studio/[organization]/clips/[stageId]/sidebar/index.tsx
