+++ packages/app/app/studio/[organization]/clips/[stageId]/sidebar/markers/DeleteMarkerButton.tsx
