+++ packages/app/app/studio/[organization]/clips/components/TimeSetter.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
-import { useClipContext } from './ClipContext';
+import { useClipContext } from '../[stageId]/ClipContext';
 import { Badge } from '@/components/ui/badge';
 
 type TimeSetterProps = {
